pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
packages = ["src/goatcounter_mcp_server"]

[project]
name = "goatcounter-mcp-server"
version = "0.1.0"
description = "Goatcounter API tool for FastMCP"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Rafal Janicki", email = "rafal@kult.io" }
]
dependencies = [
    "fastmcp>=2.2.6",
    "python-dotenv>=1.1.0",
    "httpx>=0.28.1",
    "pydantic>=2.11.4",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/rafaljanicki/goatcounter-mcp-server"
Issues = "https://github.com/rafaljanicki/goatcounter-mcp-server/issues"

[project.scripts]
goatcounter-mcp-server = "goatcounter_mcp_server:main.run"